fixes script type chip ui
