+++ apps/coordinator/.eslintrc.cjs
@@ -1,9 +1,3 @@
 module.exports = {
-  extends: [
-    "@caravan/eslint-config/app.js",
-    "plugin:react-hooks/recommended"
-  ],
-  plugins: [
-    "react-hooks"
-  ],
+  extends: ["@caravan/eslint-config/app.js"],
 };
