+++ apps/coordinator/package.json
@@ -79,7 +79,7 @@
     "test": "__GIT_SHA__=`git rev-parse --short HEAD` jest",
     "test:watch": "__GIT_SHA__=`git rev-parse --short HEAD` jest --watch",
     "test:ci": "CI=true jest --coverage",
-    "dev": " __GIT_SHA__=`git rev-parse --short HEAD` vite dev",
+    "dev": "  _GIT_SHA__=`git rev-parse --short HEAD` vite ",
     "preview": "__GIT_SHA__=`git rev-parse --short HEAD` vite preview"
   },
   "husky": {
